% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sdsbrowser.R
\name{sdsbrowser}
\alias{sdsbrowser}
\title{sdsbrowser app}
\usage{
sdsbrowser(run_app = TRUE, port = 3838, launch.browser = FALSE, ...)
}
\arguments{
\item{run_app}{Boolean. Should the app be run (TRUE - Default) or a shiny app object be returned?}

\item{port}{Integer. The TCP port that the application should listen on. Default: 2779.}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is
started. Default: FALSE.}

\item{...}{Additional arguments are passed to \code{shiny::runApp}.}
}
\description{
This function starts the sdsbrowser app with \code{shiny::runApp()}. Within this function the main user interface
is built and the different modules are loaded. All arguments are passed to \code{shiny::runApp()}, for \code{port}
and \code{launch.browser} there are defined defaults.
}
\examples{

\dontrun{
sdsbrowser::sdsbrowser()
}

}
