% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_sdsbrowser.R
\name{sdsbrowser}
\alias{sdsbrowser}
\title{sdsbrowser app}
\usage{
sdsbrowser(run_app = TRUE)
}
\arguments{
\item{run_app}{Boolean. Should the app be run (TRUE - Default) or a shiny app object be returned?}
}
\description{
sdsbrowser app
}
